Add use client declarations in shadcn components
React Server Components and Next.js believe that components
are server components by default. Some components cannot be used
on the server and have to be marked as client components
for compatibility with Next.js.

Add "use client" annotations to all components which had "use client"
in the original shadcn definition.

Reported-By: Lindenes <dodigrams@gmail.com>